// src/app/types/path.enum.ts
export enum Path {
  Home = "/",
  Console = "/console",
  Login = "/login",
  ForgotPassword = "/forgot-password",
  SecurityVerification = "/security-verification",
  ResetTwoFactor = "/reset-two-factor",
  Application = "/console/application",
  CreateApplication = "/console/application/create",
  EditApplication = "/console/application/edit",
  DetailApplication = "/console/application/edit/details",
  IntegrationApplication = "/console/application/edit/integrations",
  SecurityApplication = "/console/application/edit/security",
  Users = "/console/users",
  Roles = "/console/roles",
  Settings = "/console/settings",
  Account = "/console/settings/account",
  Project = "/console/settings/project",
  Category = "/console/settings/category",
  SubCategory = "/console/settings/sub-category",
  DashBoard = "/dashboard",
  Issue = "/console/issue",
  UserManagement = "/console/settings/user-management",
  SlaManagement = "/console/settings/sla-management",
  IssueDetail = "/console/issue/detail",
  Profile = "/console/profile",
}